1067. Digit Count in Range

+++ AllInOne/src/Math/NumberOfDigitOne.java
@@ -13,9 +13,9 @@ class Solution {
         
         int count = 0;
         
-        for (int base = 1; base <= n; base = base * 10) {
+        for (long base = 1; base <= n; base = base * 10) {
             
-            int dividend = base * 10;
+            long dividend = base * 10;
             
             // max: in case that Math.max((n % dividend) - base + 1 < 0, for example, 107 % 100 - 10 + 1 < 0
             // min: for example, 27 % 100 - 10 + 1 = 28, which is not right, it should be 10 (10 ... 19)
